+++ script/core/diagnostics/redundant-parameter.lua
@@ -32,11 +32,56 @@ local function countFuncArgs(source)
     return result
 end
 
+local function countOverLoadArgs(source, doc)
+    local result = 0
+    if source.parent and source.parent.type == 'setmethod' then
+        result = result + 1
+    end
+    local func = doc.overload
+    if not func.args or #func.args == 0 then
+        return result
+    end
+    if func.args[#func.args].type == '...' then
+        return math.maxinteger
+    end
+    result = result + #func.args
+    return result
+end
+
+local function getFuncArgs(func)
+    local funcArgs
+    local defs = vm.getDefs(func)
+    for _, def in ipairs(defs) do
+        if def.value then
+            def = def.value
+        end
+        if def.type == 'function' then
+            local args = countFuncArgs(def)
+            if not funcArgs or args > funcArgs then
+                funcArgs = args
+            end
+            if def.bindDocs then
+                for _, doc in ipairs(def.bindDocs) do
+                    if doc.type == 'doc.overload' then
+                        args = countOverLoadArgs(def, doc)
+                        if not funcArgs or args > funcArgs then
+                            funcArgs = args
+                        end
+                    end
+                end
+            end
+        end
+    end
+    return funcArgs
+end
+
 return function (uri, callback)
     local ast = files.getAst(uri)
     if not ast then
         return
     end
+
+    local cache = vm.getCache 'redundant-parameter'
 
     guide.eachSourceType(ast.ast, 'call', function (source)
         local callArgs = countCallArgs(source)
@@ -45,17 +90,12 @@ return function (uri, callback)
         end
 
         local func = source.node
-        local funcArgs
-        local defs = vm.getDefs(func)
-        for _, def in ipairs(defs) do
-            if def.value then
-                def = def.value
-            end
-            if def.type == 'function' then
-                local args = countFuncArgs(def)
-                if not funcArgs or args > funcArgs then
-                    funcArgs = args
-                end
+        local funcArgs = cache[func]
+        if funcArgs == nil then
+            funcArgs = getFuncArgs(func) or false
+            local refs = vm.getRefs(func, 0)
+            for _, ref in ipairs(refs) do
+                cache[ref] = funcArgs
             end
         end