+++ script/parser/luadoc.lua
@@ -56,6 +56,7 @@ Symbol              <-  ({} {
                         /   '...'
                         /   '+'
                         /   '#'
+                        /   '`'
                         } {})
                     ->  Symbol
 ]], {
@@ -256,6 +257,7 @@ local function parseTypeUnitArray(node)
         finish = getFinish(),
         node   = node,
     }
+    node.parent = result
     return result
 end
 
@@ -450,11 +452,40 @@ function parseType(parent)
         if not tp then
             break
         end
+
+        -- 处理 `T` 的情况
+        local typeLiteral = nil
+        if tp == 'symbol' and content == '`' then
+            nextToken()
+            if not checkToken('symbol', '`', 2) then
+                break
+            end
+            tp, content = peekToken()
+            if not tp then
+                break
+            end
+            -- TypeLiteral，指代类型的字面值。比如，对于类 Cat 来说，它的 TypeLiteral 是 "Cat"
+            typeLiteral = {
+                type   = 'doc.type.typeliteral',
+                parent = result,
+                start  = getStart(),
+                finish = nil,
+                node   = nil,
+            }
+        end
+
         if tp == 'name' then
             nextToken()
             local typeUnit = parseTypeUnit(result, content)
             if not typeUnit then
                 break
+            end
+            if typeLiteral then
+                nextToken()
+                typeLiteral.finish = getFinish()
+                typeLiteral.node = typeUnit
+                typeUnit.parent = typeLiteral
+                typeUnit = typeLiteral
             end
             result.types[#result.types+1] = typeUnit
             if not result.start then
@@ -884,6 +915,7 @@ local function buildLuaDoc(comment)
             type    = 'doc.comment',
             start   = comment.start,
             finish  = comment.finish,
+            range   = comment.finish,
             comment = comment,
         }
     end
@@ -893,6 +925,7 @@ local function buildLuaDoc(comment)
     parseTokens(doc, comment.start + startPos - 1)
     local result = convertTokens()
     if result then
+        result.range = comment.finish
         local cstart = text:find('%S', result.finish - comment.start + 2)
         if cstart and cstart < comment.finish then
             result.comment = {
@@ -904,7 +937,17 @@ local function buildLuaDoc(comment)
         end
     end
 
-    return result
+    if result then
+        return result
+    end
+
+    return {
+        type    = 'doc.comment',
+        start   = comment.start,
+        finish  = comment.finish,
+        range   = comment.finish,
+        comment = comment,
+    }
 end
 
 ---当前行在注释doc前是否有代码