+++ test.lua
@@ -7,8 +7,10 @@ package.path  = package.path
       .. ';' .. rootPath .. '/test/?/init.lua'
 local fs = require 'bee.filesystem'
 ROOT = fs.path(rootPath)
-LANG = 'zh-CN'
 TEST = true
+DEVELOP = true
+LOGPATH  = LOGPATH  or (ROOT .. '/log')
+METAPATH = METAPATH or (ROOT .. '/meta')
 
 collectgarbage 'generational'
 
@@ -18,6 +20,7 @@ log.debug('测试开始')
 ac = {}
 
 --dofile((ROOT / 'build_package.lua'):string())
+require 'tracy'
 
 local function loadAllLibs()
     assert(require 'bee.filesystem')
@@ -46,6 +49,7 @@ local function main()
     debug.setcstacklimit(1000)
     require 'parser.guide'.debugMode = true
     require 'language' 'zh-cn'
+    require 'utility'.enableCloseFunction()
     local function test(name)
         local clock = os.clock()
         print(('测试[%s]...'):format(name))
@@ -80,3 +84,5 @@ loadAllLibs()
 main()
 
 log.debug('测试完成')
+require 'bee.thread'.sleep(1)
+os.exit()